These sets of changes will rework how we handle integrations.
Instead of using the go plugin module, which is awesome, but a bit unstable
because of how we have to link and build with the exact same go version, GOPATH,
etc.

So now we're simplify going to build the server bits directly into each integrtation
as part of the agent tool chain. We can then just run the integration directly.

In the next set of changes, we'll make it such that there's a server command that will
load, run and manage each of the integrations. This will act as a control plane and the
integrations will act as the data plane.